requirements.txt
python-dotenv>=1.0.0
langchain-openai>=0.0.2
langchain-core>=0.1.0
langgraph>=0.0.15
langchain>=0.1.0
tavily-python>=0.2.8
pydantic>=2.0.0
fpdf>=1.7.2
typing-extensions>=4.5.0
numpy==1.26.3
streamlit==1.30.0
tensorflow==2.15.0
scikit-learn
firebase-admin